--- a/queries.sql
+++ b/queries.sql
@@ -175,4 +175,11 @@ GROUP BY s.name
 ORDER BY number_of_visits DESC
 LIMIT 1;
 
-
+--Performance Audit
+explain analyze SELECT COUNT(*) FROM visits where animal_id = 4
+/*
+explain analyze SELECT * FROM visits where vet_id = 2;
+run new query with temporary table
+*/
+explain analyze SELECT * FROM temp_visits;
+explain analyze SELECT * FROM owners where email = 'owner_18327@mail.com';