--- a/schema.sql
+++ b/schema.sql
@@ -57,3 +57,10 @@ CREATE TABLE visits (
   FOREIGN KEY (vet_id) REFERENCES vets (id)
 );
 
+
+--Performance Audit
+/* Add an email column to your owners table */
+ALTER TABLE owners ADD COLUMN email VARCHAR(120);
+
+
+