### Project Requirements - Database Performance Audit
> Pair Programming ( @JuanDiegoGuerra & @lerfast ) 👨🏼‍🚀

- The following queries are taking too much time (1 sec = 1000ms can be considered as too much time for database query). Try them on your machine to confirm it:

   - `SELECT COUNT(*) FROM visits where animal_id = 4;` - 1173ms🛠
![image](https://github.com/lerfast/vet_clinic_database/assets/45726631/6e0f074f-d397-4a07-bdaf-5f616a78c968)
   - `SELECT * FROM visits where vet_id = 2;` - 2288ms 🛠
![image](https://github.com/lerfast/vet_clinic_database/assets/45726631/85719cfa-9f46-4a02-bb95-3517f4dad8b9)
   - `SELECT * FROM owners where email = 'owner_18327@mail.com';` - 994ms 🛠
![image](https://github.com/lerfast/vet_clinic_database/assets/45726631/8da65557-e4aa-407f-9eab-08aec74976da)
--- 
- Result of the queries after being optimazed : 
   - `SELECT COUNT(*) FROM visits where animal_id = 4;` - 141ms ✅
![image](https://github.com/lerfast/vet_clinic_database/assets/45726631/0cf74c98-8a1c-413a-85d6-dbd2cc6b1cd3)
   - `SELECT * FROM visits where vet_id = 2;` - 955ms ✅ 
![image](https://github.com/lerfast/vet_clinic_database/assets/45726631/692df0e2-fc2d-4906-b600-7d92ba4763a4)
   - `SELECT * FROM owners where email = 'owner_18327@mail.com';` - 0.06ms ✅
![image](https://github.com/lerfast/vet_clinic_database/assets/45726631/39e514c1-3985-4b9e-8864-8efa89198d69)